feat(dependency_getter/setuptools): move logic to `PEP621DependencyGetter`
