+++ python/deptry/dependency_getter/builder.py
@@ -53,16 +53,6 @@ class DependencyGetterBuilder:
             if self._project_uses_uv(pyproject_toml):
                 return UvDependencyGetter(self.config, self.package_module_name_map, self.pep621_dev_dependency_groups)
 
-            if self._project_uses_setuptools(pyproject_toml):
-                project_uses_dynamic_dependencies, requirements_files, requirements_files_dev = (
-                    self._project_uses_setuptools_dynamic_dependencies(pyproject_toml)
-                )
-
-                if project_uses_dynamic_dependencies:
-                    return RequirementsTxtDependencyGetter(
-                        self.config, self.package_module_name_map, requirements_files, requirements_files_dev
-                    )
-
             if self._project_uses_pep_621(pyproject_toml):
                 return PEP621DependencyGetter(
                     self.config, self.package_module_name_map, self.pep621_dev_dependency_groups
@@ -134,63 +124,6 @@ class DependencyGetterBuilder:
             return False
         else:
             return True
-
-    @staticmethod
-    def _project_uses_setuptools(pyproject_toml: dict[str, Any]) -> bool:
-        try:
-            if pyproject_toml["build-system"]["build-backend"] == "setuptools.build_meta":
-                logging.debug(
-                    "pyproject.toml has the entry build-system.build-backend == 'setuptools.build_meta', so setuptools is used to specify the project's dependencies."
-                )
-                return True
-            else:
-                logging.debug(
-                    "pyproject.toml does not have build-system.build-backend == 'setuptools.build_meta', so setuptools "
-                    "is not used to specify the project's dependencies."
-                )
-                return False
-        except KeyError:
-            logging.debug(
-                "pyproject.toml does not contain a build-system.build-backend entry, so setuptools is not used to "
-                "specify the project's dependencies."
-            )
-            return False
-
-    def _project_uses_setuptools_dynamic_dependencies(
-        self, pyproject_toml: dict[str, Any]
-    ) -> tuple[bool, tuple[str, ...], tuple[str, ...]]:
-        try:
-            if "dependencies" not in pyproject_toml["project"]["dynamic"]:
-                logging.debug(
-                    "pyproject.toml does not have dependencies listed as dynamic metadata, so dynamic dependencies are "
-                    "not used."
-                )
-            else:
-                pyproject_toml["tool"]["setuptools"]["dynamic"]["dependencies"]["file"]
-                logging.debug(
-                    "pyproject.toml has dependencies listed as dynamic metadata and contains a populated "
-                    "tool.setuptools.dynamic.dependencies.file entry, so dynamic dependencies are used."
-                )
-
-                requirements_files = pyproject_toml["tool"]["setuptools"]["dynamic"]["dependencies"]["file"]
-                requirements_files_dev = []
-
-                for group, specification in (
-                    pyproject_toml["tool"]["setuptools"]["dynamic"].get("optional-dependencies", {}).items()
-                ):
-                    if group in self.pep621_dev_dependency_groups:
-                        requirements_files_dev.extend(specification["file"])
-                    else:
-                        requirements_files.extend(specification["file"])
-
-                return True, tuple(requirements_files), tuple(requirements_files_dev)
-        except KeyError:
-            logging.debug(
-                "pyproject.toml either does not contain a project.dynamic entry or "
-                "tool.setuptools.dynamic.dependencies.file entry, so dynamic dependencies are not used."
-            )
-
-        return False, (), ()
 
     @staticmethod
     def _project_uses_pep_621(pyproject_toml: dict[str, Any]) -> bool:
